Added function convertDay and additions to the parseFile function

+++ icsParser.py
@@ -1,7 +1,7 @@
 import sys
 import re
 import shutil
-import os
+import os.path
 from datetime import datetime
 
 class Event:
@@ -60,6 +60,7 @@ class Parser:
         tempRawTime   = ''
         tempFreq      = ''
         tempRREND     = ''
+        tempRRDAY     = ''
 
         for line in lines:
             # Find an Event
@@ -71,12 +72,14 @@ class Parser:
             if(any(identifier in line for identifier in self.EVENT_START_TIME) and eventFound):
                 # Strip date/time from string
                 time = re.sub('[^0-9]','',line)
-                # Fix for times earlier than 10 not appearing in the schedule
-                # due to indexing error
-                if(int(time[8:10]) < 10):
-                    tempStartTime = time[9:10]
-                else:
-                    tempStartTime = time[8:10]
+                # Check to see if time is in the correct format
+                if(int(time) > 20000000000000): 
+                    # Fix for times earlier than 10 not appearing in the schedule
+                    # due to indexing error
+                    if(int(time[8:10]) < 10):
+                        tempStartTime = time[9:10]
+                    else:
+                        tempStartTime = time[8:10]
                 continue
 
             # Get event end time
@@ -107,6 +110,10 @@ class Parser:
                             tempRREND = temp
                 elif("FREQ=DAILY" in line):
                     tempFreq  = "DAILY"
+                    # tempRRDAY would be null when events of daily reoccurance
+                    # this will use the rawTime variable to calculate what week day the event repeats on
+                    dt        = datetime.strptime(str(tempRawTime)+ " 00:00:00",'%Y%m%d %H:%M:%S')
+                    tempRRDAY = self.convertDay(dt.weekday())
                     if("UNTIL=" not in line):
                         tempRREND = -1
                     # Get datetime and convert into an int of the event
@@ -151,6 +158,24 @@ class Parser:
         elif re.search(r'\b' + "BYDAY=SU"+ r'\b',line):
             return "BYDAY=SU"
     
+    # Used to find what day of the week Daily events start on by converting a datetime to one of the 
+    # return values bellow
+    def convertDay(self,val):
+        if (val == 0):
+            return "BYDAY=MO"
+        elif (val == 1):
+            return "BYDAY=TU"
+        elif (val == 2):
+            return "BYDAY=WE"
+        elif (val == 3):
+            return "BYDAY=TH"
+        elif (val == 4):
+            return "BYDAY=FR"
+        elif (val == 5):
+            return "BYDAY=SA"
+        elif (val == 6):
+            return "BYDAY=SU"
+
     # This will create a new event list with only newer events
     def cleanList(self):
         # Loop through EVENT_LIST
@@ -207,19 +232,34 @@ class Parser:
     def printEvents(self):
         # Remove old events
         self.cleanList()
-        try:
-            with open('parseFile.txt','w',encoding='utf-8',errors='ignore') as file:
+        if(os.path.isfile("./scripts/parseFile.txt")):
+            with open('./scripts/parseFile.txt','a',encoding='utf-8',errors='ignore') as file:
                 sys.stdout = file
                 for i,evnt in enumerate(self.CLEAN_EVENT_LIST):
-                    print("-----EVENT " + str(i+1))
-                    print(evnt.name)  
-                    print(evnt.startTime)
-                    print(evnt.rawDate)
-                    print(evnt.rruleDay)
-                # Moves new parse file to the scripts folder
-                shutil.move("./parseFile.txt","./scripts/parseFile.txt")
-        except OSError:
-            print("Something went wrong writing to parse file!")
+                    file.write("-----EVENT " + str(i+1))
+                    file.write("\n") 
+                    file.write(evnt.name) 
+                    file.write("\n")
+                    file.write(evnt.startTime)
+                    file.write("\n") 
+                    file.write(str(evnt.rawDate))
+                    file.write("\n") 
+                    file.write(str(evnt.rruleDay))
+                    file.write("\n") 
+        else:
+            try:
+                with open('parseFile.txt','w',encoding='utf-8',errors='ignore') as file:
+                    sys.stdout = file
+                    for i,evnt in enumerate(self.CLEAN_EVENT_LIST):
+                        print("-----EVENT " + str(i+1))
+                        print(evnt.name)  
+                        print(evnt.startTime)
+                        print(evnt.rawDate)
+                        print(evnt.rruleDay)
+                    # Moves new parse file to the scripts folder
+                    shutil.move("./parseFile.txt","./scripts/parseFile.txt")
+            except OSError:
+                print("Something went wrong writing to parse file!")
 p = Parser()
 p.parseFile()
 p.printEvents()
